--- src/commands_pre_player.c
@@ -1,34 +1,0 @@
-#include "serveur.h"
-
-void	command_pre_avance(t_zappy *var, t_player *p, char *args)
-{
-	(void)args;
-	// send message to gfx_client
-	// add player action to action queue
-}
-
-void	command_pre_droite(t_zappy *var, t_player *p, char *args)
-{
-	(void)args;
-	// send message to gfx_client
-	// add player action to action queue
-}
-
-void	command_pre_gauche(t_zappy *var, t_player *p, char *args)
-{
-	(void)args;
-	// send message to gfx_client
-	// add player action to action queue
-}
-
-void	command_pre_voir(t_zappy *var, t_player *p, char *args)
-{
-	(void)args;
-	// add player action to action queue
-}
-
-void	command_pre_inventaire(t_zappy *var, t_player *p, char *args)
-{
-	(void)args;
-	// add player action to action queue
-}
