+++ src/serveur.h
@@ -40,6 +40,38 @@
 # include "structs.h"
 
 /*
+** src/action_player.c
+*/
+void	command_player_move(t_zappy *var, t_player *p, int dir);
+void	command_player_avance(t_zappy *var, t_player *p, char *args);
+void	command_player_droite(t_zappy *var, t_player *p, char *args);
+void	command_player_gauche(t_zappy *var, t_player *p, char *args);
+void	command_player_expulse(t_zappy *var, t_player *p, char *args);
+
+/*
+** src/action_player_2.c
+*/
+void	command_player_voir_sub(t_zappy *var, t_player *p, int k, int l);
+void	command_player_voir(t_zappy *var, t_player *p, char *args);
+void	command_player_inventaire(t_zappy *var, t_player *p, char *args);
+void	command_player_prend(t_zappy *var, t_player *p, char *args);
+void	command_player_pose(t_zappy *var, t_player *p, char *args);
+
+/*
+** src/action_player_3.c
+*/
+void	command_player_incantation(t_zappy *var, t_player *p, char *args);
+void	command_player_fork(t_zappy *var, t_player *p, char args);
+
+/*
+** src/action_player_broadcast.c
+*/
+int		broadcast_get_distance(int ms[2], int s[2], int r[2], int i);
+int		broadcast_get_direction(int ms[2], t_player *s, t_player *r);
+int		broadcast_get_square(int ms[2], t_player *s, t_player *r);
+void	command_player_broadcast(t_zappy *var, t_player *p, char *args);
+
+/*
 ** src/check_arguments.c
 */
 int			check_arguments(t_arguments *args, int error);
